+++ Zadanie1/3/README.md
@@ -1,4 +1,4 @@
-* [ ] 1 rysunek dla zmian błędu średniokwadratowego dla warstwy ukrytej z 1, 2 i 3 neuronami bez biasu (learning rate = 0.1, bez momentum)
-* [ ] to samo ale z biasem
+* [x] 1 rysunek dla zmian błędu średniokwadratowego dla warstwy ukrytej z 1, 2 i 3 neuronami bez biasu (learning rate = 0.1, bez momentum)
+* [x] to samo ale z biasem
 * [ ] tabela prezentująca uśredniony czas nauki (liczba iteracji) potrzebnej dla osiągnięcia zadanej wartości błędu średniokwadratowego oraz odchylenia standardowego ze 100 prób (każda próba różni się od innych początkowymi wartościami wag) dla różnych kombinacji parametrów nauki (współczynnik nauki i momentum), należy przyjąć architekturę z 2 neuronami w warstwie ukrytej i biasem
 * [ ] 2 neurony w warstwie ukrytej, stałe parametry nauki, dwie tabele (jedna z biasem druga bez biasu) prezentujące odpowiedzi neurowów z warstwy ukrytej dla każdej danej wejściowej z zestawu treningowego
