+++ Zadanie1/Functions.py
@@ -1,18 +1,18 @@
 ''' Common functions '''
-import csv
+import numpy as np
 
 
 #Mean squared error
 def MSE(f,intput,output):
     ans=0;
     for i in range(len(output)):
-        ans+=((f(intput[i])-output[i])**2)
-    return ans/2 
+        ans+=((f(intput[i]).T-output[i])**2)
+    return np.sum(ans)/2
 
-def getData(intput):
-    reader = csv.reader(intput)
-    outX=[]
-    for row in reader:
-        i = list(map(float,row[0].split(" ")))
-        outX.append(i)
-    return outX
+#Custom ticks to plot functions 
+def myticks(x,pos):
+    if x <= 0:
+         return "$0$"
+    exponent = int(np.log10(x))
+    coeff = x/10**exponent
+    return r"${:2.1f} \times 10^{{ {:2d} }}$".format(coeff,exponent)
